[Silver III] Title: 주유소, Time: 952 ms, Memory: 186208 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -2,27 +2,27 @@ import java.util.*;
 public class Main {
 
     static int N;
-    static List<Integer> dist = new ArrayList<>();
-    static List<Integer> litter = new ArrayList<>();
+    static List<Long> dist = new ArrayList<>();
+    static List<Long> litter = new ArrayList<>();
 
     public static void main(String[] args) {
 
         Scanner in = new Scanner(System.in);
         N = in.nextInt();
         for(int i=0;i<N-1;i++) {
-            dist.add(in.nextInt());
+            dist.add(in.nextLong());
         }
         for(int i=0;i<N;i++) {
-            litter.add(in.nextInt());
+            litter.add(in.nextLong());
         }
 
-        int ret=0;
-        int min_Litter=987654321;
+        long ret=0;
+        long min_Litter=987654321;
         for(int i=0;i<N-1;i++) {
             min_Litter=Math.min(min_Litter,litter.get(i));
             ret+=min_Litter*dist.get(i);
         }
-        
+
         System.out.println(ret);
     }
 }
